requirements.txt
absl-py==1.4.0
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
boto3==1.26.76
botocore==1.29.76
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.7
croniter==1.3.14
cycler==0.11.0
dateutils==0.6.12
deepdiff==6.3.0
Deprecated==1.2.13
dicom2nifti==2.4.8
dill==0.3.6
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.6.0
fastapi==0.88.0
filelock==3.11.0
flash-attn==1.0.1
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.53.0
h11==0.14.0
humanize==4.6.0
idna==3.4
imageio==2.27.0
importlib-resources==5.12.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonargparse==4.20.1
kiwisolver==1.4.4
lazy_loader==0.2
lightning==2.0.1.post0
lightning-cloud==0.5.33
lightning-utilities==0.8.0
lit==16.0.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
monai==1.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
nibabel==5.1.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
ordered-set==4.1.0
p-tqdm==1.4.0
packaging==23.1
pandas==2.0.0
pathlib==1.0.1
pathos==0.3.0
pathtools==0.1.2
Pillow==9.5.0
pox==0.3.2
ppft==1.7.6.6
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
pydicom==2.3.1
Pygments==2.15.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-gdcm==3.0.21
python-multipart==0.0.6
pytorch-lightning==2.0.1.post0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
readchar==4.0.5
requests==2.28.2
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
s3fs==2023.4.0
s3transfer==0.6.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.19.1
setproctitle==1.3.2
shellingham==1.5.0.post1
SimpleITK==2.2.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
threadpoolctl==3.1.0
tifffile==2023.4.12
torchmetrics==0.11.4
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typer==0.7.0
typeshed-client==2.2.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
wandb==0.14.2
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.1
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
